Spawn movable particles. Added nucleus(class Atom) which substitutes strong interactions and helps dealing with radioactive decay etc. Added decay beta(-) and neutron outburst. Tweaked forces a bit.

+++ Atomizer/src/atomizer/particles/Electron.java
@@ -11,7 +11,7 @@ import atomizer.Vector;
  * @author radek
  */
 public class Electron extends Particle {
-    public Electron(int x,int y,Vector speed){
+    public Electron(double x,double y,Vector speed){
         super(x,y,speed);
         mass = 1.0/50.0;//too little for scale
         charge = -1.0;
